`data:api`、`data:repository`モジュールを追加
