+++ app/src/debug/java/org/ageage/eggplant/DebugApp.kt
@@ -1,7 +1,7 @@
 package org.ageage.eggplant
 
 import android.app.Application
-import org.ageage.eggplant.common.api.Client
+import org.ageage.eggplant.data.api.Client
 import timber.log.Timber
 
 class DebugApp : Application() {
