+++ app/src/main/java/org/ageage/eggplant/App.kt
@@ -1,7 +1,7 @@
 package org.ageage.eggplant
 
 import android.app.Application
-import org.ageage.eggplant.common.api.Client
+import org.ageage.eggplant.data.api.Client
 
 class App : Application() {
     override fun onCreate() {
