--- app/src/main/java/org/ageage/eggplant/common/api/response/BookmarkEntryResponse.kt
@@ -1,12 +1,0 @@
-package org.ageage.eggplant.common.api.response
-
-import com.google.gson.annotations.SerializedName
-
-data class BookmarkEntryResponse(
-    val title: String,
-    val count: Int,
-    val entryUrl: String,
-    val screenshot: String,
-    val eid: String,
-    @SerializedName("bookmarks") val bookmarkResponses: List<BookmarkResponse>
-)
