--- app/src/main/java/org/ageage/eggplant/common/api/response/BookmarkStarResponse.kt
@@ -1,8 +1,0 @@
-package org.ageage.eggplant.common.api.response
-
-import com.google.gson.annotations.SerializedName
-
-data class BookmarkStarResponse(
-    val entries: List<BookmarkStarEntry>,
-    @SerializedName("can_comment") val canComment: String
-)
