--- app/src/main/java/org/ageage/eggplant/common/api/response/ColoredStar.kt
@@ -1,6 +1,0 @@
-package org.ageage.eggplant.common.api.response
-
-data class ColoredStar(
-    val stars: List<Star>,
-    val color: String
-)
