--- app/src/main/java/org/ageage/eggplant/common/api/response/Star.kt
@@ -1,6 +1,0 @@
-package org.ageage.eggplant.common.api.response
-
-data class Star(
-    val quote: String,
-    val name: String
-)
