--- app/src/main/java/org/ageage/eggplant/common/enums/Endpoint.kt
@@ -1,12 +1,0 @@
-package org.ageage.eggplant.common.enums
-
-enum class Endpoint(
-    val url: String
-) {
-    HATENA_BOOKMARK(
-        "https://b.hatena.ne.jp"
-    ),
-    HATENA_STAR(
-        "https://s.hatena.com"
-    )
-}
