+++ app/src/test/java/org/ageage/eggplant/feed/FeedItemsViewModelTest.kt
@@ -7,10 +7,10 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
-import org.ageage.eggplant.common.api.response.Item
 import org.ageage.eggplant.common.enums.Category
 import org.ageage.eggplant.common.enums.Mode
-import org.ageage.eggplant.common.repository.FeedRepository
+import org.ageage.eggplant.data.api.response.Item
+import org.ageage.eggplant.data.repository.FeedRepository
 import org.junit.After
 import org.junit.Before
 import org.junit.Rule
