+++ app/src/test/java/org/ageage/eggplant/search/SearchResultViewModelTest.kt
@@ -7,12 +7,12 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.runBlocking
 import kotlinx.coroutines.test.resetMain
 import kotlinx.coroutines.test.setMain
-import org.ageage.eggplant.common.api.response.Item
 import org.ageage.eggplant.common.enums.MinimumBookmarkCount
 import org.ageage.eggplant.common.enums.SearchTarget
 import org.ageage.eggplant.common.enums.SortType
 import org.ageage.eggplant.common.model.SearchFilterOption
-import org.ageage.eggplant.common.repository.FeedRepository
+import org.ageage.eggplant.data.api.response.Item
+import org.ageage.eggplant.data.repository.FeedRepository
 import org.assertj.core.api.AssertionsForClassTypes.assertThat
 import org.junit.After
 import org.junit.Before
