+++ data/api/src/androidTest/java/org/ageage/eggplant/data/api/ExampleInstrumentedTest.kt
@@ -1,0 +1,22 @@
+package org.ageage.eggplant.data.api
+
+import androidx.test.platform.app.InstrumentationRegistry
+import androidx.test.runner.AndroidJUnit4
+import org.junit.Assert.assertEquals
+import org.junit.Test
+import org.junit.runner.RunWith
+
+/**
+ * Instrumented test, which will execute on an Android device.
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+@RunWith(AndroidJUnit4::class)
+class ExampleInstrumentedTest {
+    @Test
+    fun useAppContext() {
+        // Context of the app under test.
+        val appContext = InstrumentationRegistry.getInstrumentation().targetContext
+        assertEquals("org.ageage.eggplant.data.api.test", appContext.packageName)
+    }
+}
