+++ data/api/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="org.ageage.eggplant.data.api" />
