+++ data/api/src/main/java/org/ageage/eggplant/data/api/response/BookmarkResponse.kt
@@ -1,0 +1,9 @@
+package org.ageage.eggplant.data.api.response
+
+data class BookmarkResponse(
+    val user: String,
+    val tags: List<String>,
+    val timestamp: String,
+    val comment: String,
+    var entry: BookmarkStarEntry? = null
+)
