+++ data/api/src/main/java/org/ageage/eggplant/data/api/response/BookmarkStarResponse.kt
@@ -1,0 +1,8 @@
+package org.ageage.eggplant.data.api.response
+
+import com.google.gson.annotations.SerializedName
+
+data class BookmarkStarResponse(
+    val entries: List<BookmarkStarEntry>,
+    @SerializedName("can_comment") val canComment: String
+)
