+++ data/api/src/main/java/org/ageage/eggplant/data/api/response/ColoredStar.kt
@@ -1,0 +1,6 @@
+package org.ageage.eggplant.data.api.response
+
+data class ColoredStar(
+    val stars: List<Star>,
+    val color: String
+)
