+++ data/api/src/main/java/org/ageage/eggplant/data/api/response/Item.kt
@@ -1,0 +1,10 @@
+package org.ageage.eggplant.data.api.response
+
+data class Item(
+    var title: String,
+    var link: String,
+    var description: String,
+    var bookmarkCount: String,
+    var imageUrl: String,
+    var faviconUrl: String
+)
