+++ data/api/src/main/java/org/ageage/eggplant/data/api/response/Star.kt
@@ -1,0 +1,6 @@
+package org.ageage.eggplant.data.api.response
+
+data class Star(
+    val quote: String,
+    val name: String
+)
