+++ data/repository/src/main/java/org/ageage/eggplant/data/repository/model/Bookmark.kt
@@ -1,0 +1,11 @@
+package org.ageage.eggplant.data.repository.model
+
+import java.util.*
+
+data class Bookmark(
+    val user: String,
+    val tags: List<String>,
+    val timeStamp: Date,
+    val comment: String,
+    val yellowStarNumber: Int
+)
