+++ settings.gradle
@@ -1,1 +1,1 @@
-include ':app'
+include ':app', 'data:repository', 'data:api'
